backend/backend/requirements.txt
asgiref==3.9.1
black==24.1.1
click==8.2.1
colorama==0.4.6
coverage==7.10.5
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
flake8==7.0.0
gunicorn==23.0.0
iniconfig==2.1.0
isort==5.13.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
python-dotenv==1.1.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
